pyproject.toml
[project]
name = "astro-tui"
version = "0.1.0"
description = "Terminal chatbot showcasing the astro-context library"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Arthur cavalcanti Granja", email = "artcgranja@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "astro-context[all]",
    "textual>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
astro-tui = "astro_tui.app:main"

[tool.uv.sources]
astro-context = { path = "../astro-context", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
